Move helper functions to a separate module
